record button in own component
